NumberResponse

+++ app/src/main/java/com/example/randomnumber/models/NumberResponse.kt
@@ -1,9 +1,10 @@
 package com.example.randomnumber.models
 
+
 data class NumberResponse(
     val text: String? = null,
     val found: Boolean? = null,
-    val number: Int? = null,
+    val number: Long? = null,
     val type: String? = null,
     val date: String? = null
 )
